Refactor file writing to use File.open and a block
Previously we were using File.new to access a file. However, if an
error occurs before we call File#close on the IO instance created,
the file isn't properly closed and the data may be corrupted?

Using open and a block ensures that whenever we leave the block,
be it for an error or exiting our code normally, it will automatically
close the file.