Reset the account before loading from file
Previously when loading from a file, we would simply use Account#deposit
and Account#withdraw to populate the account.

However, when the account already had transactions, we would simply
push the transactions from the file on to those. Not only is this
behaviour undesirable, it could raise errors if the file transactions
we not in chronological order with the ones already in the account.